Move EnumObject to shared location
