+++ src/DataView/Operator.php
@@ -2,6 +2,8 @@
 
 namespace DataKit\DataViews\DataView;
 
+use DataKit\DataViews\EnumObject;
+
 /**
  * Represents a valid filter operator.
  *
